Use SectionType for cellForRowAt:indexPath method. Extract metric calculations
